+++ db_status.go
@@ -1,101 +1,28 @@
 package main
 
 import (
-	"errors"
+	"context"
+	"time"
 
-	"gopkg.in/mgo.v2"
-	"gopkg.in/mgo.v2/bson"
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
-// SetStatus 함수는 Status를 DB에 저장한다.
-func SetStatus(session *mgo.Session, s Status) error {
-	session.SetMode(mgo.Monotonic, true)
-	c := session.DB("setting").C("status")
-	err := c.Update(bson.M{"id": s.ID}, s)
+func AllStatusV2(client *mongo.Client) ([]Status, error) {
+	collection := client.Database("setting").Collection("status")
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+	findOptions := options.Find()
+	findOptions.SetSort(bson.D{{"order", -1}})
+	var status []Status
+	cursor, err := collection.Find(ctx, bson.D{}, findOptions)
 	if err != nil {
-		if err == mgo.ErrNotFound {
-			err = c.Insert(s)
-			if err != nil {
-				return err
-			}
-			return nil
-		}
-		return err
+		return status, err
 	}
-	return nil
-}
-
-// GetStatus 함수는 Status를 DB에서 가지고 온다.
-func GetStatus(session *mgo.Session, id string) (Status, error) {
-	session.SetMode(mgo.Monotonic, true)
-	c := session.DB("setting").C("status")
-	s := Status{}
-	err := c.Find(bson.M{"id": id}).One(&s)
+	err = cursor.All(ctx, &status)
 	if err != nil {
-		return s, err
+		return status, err
 	}
-	return s, nil
-}
-
-// GetInitStatusID 함수는 초기 Status를 DB에서 가지고 온다.
-func GetInitStatusID(session *mgo.Session) (string, error) {
-	session.SetMode(mgo.Monotonic, true)
-	c := session.DB("setting").C("status")
-	s := Status{}
-	n, err := c.Find(bson.M{"initstatus": true}).Count()
-	if err != nil {
-		return "", err
-	}
-	if n == 0 {
-		return "", errors.New("아이템 생성시 초기 Status로 사용할 Status 설정이 필요합니다")
-	}
-	if n != 1 {
-		return "", errors.New("초기 상태 설정값이 1개가 아닙니다")
-	}
-	err = c.Find(bson.M{"initstatus": true}).One(&s)
-	if err != nil {
-		return "", err
-	}
-	return s.ID, nil
-}
-
-// AddStatus 함수는 tasksetting을 DB에 추가한다.
-func AddStatus(session *mgo.Session, s Status) error {
-	session.SetMode(mgo.Monotonic, true)
-	c := session.DB("setting").C("status")
-	n, err := c.Find(bson.M{"id": s.ID}).Count()
-	if err != nil {
-		return err
-	}
-	if n > 0 {
-		return errors.New(s.ID + " Status가 이미 존재합니다")
-	}
-	err = c.Insert(s)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// RmStatus 함수는 Status를 DB에서 삭제한다.
-func RmStatus(session *mgo.Session, id string) error {
-	session.SetMode(mgo.Monotonic, true)
-	c := session.DB("setting").C("status")
-	err := c.Remove(bson.M{"id": id})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// AllStatus 함수는 모든 Status값을 DB에서 가지고 온다.
-func AllStatus(session *mgo.Session) ([]Status, error) {
-	session.SetMode(mgo.Monotonic, true)
-	c := session.DB("setting").C("status")
-	results := []Status{}
-	err := c.Find(bson.M{}).Sort("-order").All(&results)
-	if err != nil {
-		return nil, err
-	}
-	return results, nil
+	return status, nil
 }
